+++ server/githubsrm/apis/models.py
@@ -1,38 +1,13 @@
-import random
-import string
 from typing import Any, Dict
 
-import pymongo
-from django.conf import settings
-from dotenv import load_dotenv
+from core.models import BaseModel
 
 from .errors import MiscErrors, ProjectErrors
 
-load_dotenv()
 
-
-class Entry:
-    def __init__(self):
-        client = pymongo.MongoClient(settings.DATABASE["mongo_uri"])
-        self.db = client[settings.DATABASE["db"]]
-
-    def get_uid(self) -> str:
-        """Returns 8 character alpha numeric unique id
-
-        Args:
-            length (int)
-            operator (pymongo.MongoClient)
-
-        Returns:
-            str
-        """
-
-        gen_id = random.choices(string.ascii_uppercase + string.digits, k=8)
-
-        if self.db.collection.find_one({"_id": gen_id}):
-            return self.get_uid(length=8)
-
-        return "".join(gen_id)
+class Entry(BaseModel):
+    def __init__(self) -> None:
+        super().__init__()
 
     def _enter_project(
         self, doc: Dict[str, str], visibility: bool, project_id: str
